rename ChatBar to Sidebar
