+++ src/App.css
@@ -218,7 +218,7 @@ footer a {
   color: #fff;
 }
 
-.chatBar {
+.sideBar {
   min-height: 100%;
   width: 20%;
   float: right;
